pyproject.toml
[tool.poetry]
name = "data_job_crawler"
version = "1.1.0"
description = ""
authors = ["Donor <donorfelita@msn.com>"]

[tool.poetry.dependencies]
python = "^3.9"
psycopg2-binary = "^2.9.5"
scrapy = "^2.6.2"
playwright = "^1.24.1"
scrapy-playwright = "*"
[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
